Configured needed script files for continuous integrating and deployment.
Currently, Gitlab mirrors our repository and refreshes it every ~15min. Whenever a commit is pushed to master, it runs a pipeline consisting of test and publish. Then resulting docker image is pushed to Gitlab Docker Registry. In our production server, Watchman always check if there is a new update on `latest` tag, if so it pulls and restarts the docker container with the new image. One push and at most in 20 minutes, it's in production.

- .gitlab-ci.yml is configured.
- Production server is configured.
- GitLab is configured.

jira SPRCOUR8-8